+++ src/clua-i-machine.cpp
@@ -911,6 +911,17 @@ static int machine_obj_index_remove_stored(lua_State *L) {
     return 0;
 }
 
+/// \brief This is the machine:sync_stored() method implementation.
+/// \param L Lua state.
+static int machine_obj_index_sync_stored(lua_State *L) {
+    auto &m = clua_check<clua_managed_cm_ptr<cm_machine>>(L, 1);
+    const char *dir = luaL_checkstring(L, 2);
+    if (cm_sync_stored(m.get(), dir) != 0) {
+        return luaL_error(L, "%s", cm_get_last_error_message());
+    }
+    return 0;
+}
+
 /// \brief This is the machine:verify_hash_tree() method implementation.
 /// \param L Lua state.
 static int machine_obj_index_verify_hash_tree(lua_State *L) {
@@ -1403,6 +1414,7 @@ static const auto machine_obj_index = cartesi::clua_make_luaL_Reg_array({
     {.name = "store", .func = machine_obj_index_store},
     {.name = "clone_stored", .func = machine_obj_index_clone_stored},
     {.name = "remove_stored", .func = machine_obj_index_remove_stored},
+    {.name = "sync_stored", .func = machine_obj_index_sync_stored},
     {.name = "swap", .func = machine_obj_index_swap},
     {.name = "translate_virtual_address", .func = machine_obj_index_translate_virtual_address},
     {.name = "verify_hash_tree", .func = machine_obj_index_verify_hash_tree},