+++ src/jsonrpc-machine.cpp
@@ -712,6 +712,11 @@ void jsonrpc_machine::do_remove_stored(const std::string &dir) const {
     request("machine.remove_stored", std::tie(dir), result);
 }
 
+void jsonrpc_machine::do_sync_stored(const std::string &dir) const {
+    bool result = false;
+    request("machine.sync_stored", std::tie(dir), result);
+}
+
 uint64_t jsonrpc_machine::do_read_reg(reg r) const {
     uint64_t result = 0;
     request("machine.read_reg", std::tie(r), result);