+++ src/jsonrpc-remote-machine.cpp
@@ -1084,6 +1084,20 @@ static json jsonrpc_machine_remove_stored_handler(const json &j, const std::shar
     return jsonrpc_response_ok(j);
 }
 
+/// \brief JSONRPC handler for the machine.sync_stored method
+/// \param j JSON request object
+/// \param session HTTP session
+/// \returns JSON response object
+static json jsonrpc_machine_sync_stored_handler(const json &j, const std::shared_ptr<http_session> &session) {
+    if (!session->handler->machine) {
+        return jsonrpc_response_invalid_request(j, "no machine");
+    }
+    static const char *const param_name[] = {"dir"};
+    auto args = parse_args<std::string>(j, param_name);
+    cartesi::machine::sync_stored(std::get<0>(args));
+    return jsonrpc_response_ok(j);
+}
+
 /// \brief JSONRPC handler for the machine.run method
 /// \param j JSON request object
 /// \param session HTTP session
@@ -1749,6 +1763,7 @@ static json jsonrpc_dispatch_method(const json &j, const std::shared_ptr<http_se
         {"machine.store", jsonrpc_machine_store_handler},
         {"machine.clone_stored", jsonrpc_machine_clone_stored_handler},
         {"machine.remove_stored", jsonrpc_machine_remove_stored_handler},
+        {"machine.sync_stored", jsonrpc_machine_sync_stored_handler},
         {"machine.run", jsonrpc_machine_run_handler},
         {"machine.log_step", jsonrpc_machine_log_step_handler},
         {"machine.run_uarch", jsonrpc_machine_run_uarch_handler},